core: Implicit narrowing conversion from long to int
Previously compound assignment resulted in an implicit narrowing conversion from long to int.
This was due to two reasons:
- An integer value got stored in a long and then later used back in a compound assignment with an integer.
- The stream count method returns a long whichs is used together with an integer value in a compound assignment.

This can result in information loss and numeric errors such as overflows.
Removed unneeded conversions where integers were stored in a long to be later be used as an integer again.
For the stream issue this is solved by altering the type of variable from int to long which saves the stream return value.
The type of variable does not matter for these instances because the variable is only used internally in the function.

+++ backend/manager/modules/bll/src/main/java/org/ovirt/engine/core/bll/scheduling/utils/VdsCpuUnitPinningHelper.java
@@ -112,7 +112,7 @@ public class VdsCpuUnitPinningHelper {
         List<VdsCpuUnit> cpusToBeAllocated = allocateExclusiveCpus(cpuTopology, vm, hostId);
         if (vm.getCpuPinningPolicy() == CpuPinningPolicy.ISOLATE_THREADS) {
             List<Integer> socketsUsed = cpusToBeAllocated.stream().map(VdsCpuUnit::getSocket).distinct().collect(Collectors.toList());
-            int pCores = 0;
+            long pCores = 0;
             for (int socket: socketsUsed) {
                 pCores += getCoresInSocket(cpusToBeAllocated, socket).stream().map(VdsCpuUnit::getCore).distinct().count();
             }
